+++ packages/hourly-chart/components/HourlyChart/index.jsx
@@ -1,14 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Text,
-  Button,
-} from '@bufferapp/components';
-
-import {
-  PlusIcon,
-} from '@bufferapp/components/Icon/Icons';
-
+import { Text } from '@bufferapp/components';
 
 import {
   geyser,
@@ -20,8 +12,7 @@ import {
 
 import PostCountByHour from '../PostCountByHour';
 import HourlyEngagementChart from '../HourlyEngagementChart';
-import TimezoneInfo from '../TimezoneInfo';
-import Dropdown from '../Dropdown';
+import ChartHeader from '../ChartHeader';
 
 const title = {
   margin: '2rem 0 1rem',
@@ -40,47 +31,14 @@ const gridContainer = {
   border: `1px solid ${geyser}`,
 };
 
-const chartHeader = {
-  display: 'flex',
-  alignItems: 'center',
-  padding: '1.25rem',
-};
-
-const toggleWrapper = {
-  marginLeft: '1.25rem',
-};
-
-const closeButton = {
-  marginTop: 'auto',
-  marginBottom: 'auto',
-  marginLeft: '10px',
-  cursor: 'pointer',
-  lineHeight: '1',
-};
-
-const openButton = {
-  display: 'inline-block',
-  padding: '7px 11px',
-  border: '1px solid #ced7df',
-  borderRadius: '3px',
-  cursor: 'pointer',
-};
-
-const SecondaryMetricToggle = ({ metric, metrics }) => (
-  <span style={toggleWrapper}>
-    { metric && <Dropdown metrics={metrics} selectedMetric={metric} />}
-    { metric && <Button noStyle><Text size="small" color="geyser"><span style={closeButton}>x</span></Text></Button>}
-    { !metric && <Button noStyle><span style={openButton}><PlusIcon size="small" /></span></Button>}
-  </span>
-);
-
 const ChartContent = ({ posts, timezone, metrics, selectedMetric, secondaryMetric }) => (
   <div>
-    <section style={chartHeader}>
-      <Dropdown metrics={metrics} selectedMetric={selectedMetric} />
-      <SecondaryMetricToggle metric={secondaryMetric} metrics={metrics} />
-      <TimezoneInfo timezone={timezone} />
-    </section>
+    <ChartHeader
+      metrics={metrics}
+      selectedMetric={selectedMetric}
+      secondaryMetric={secondaryMetric}
+      timezone={timezone}
+    />
     <HourlyEngagementChart
       posts={posts}
       metric={metrics[selectedMetric]}
