+++ packages/hourly-chart/package.json
@@ -8,7 +8,8 @@
     "lint": "eslint . --ignore-pattern coverage node_modules",
     "test": "yarn run lint && sh ../../package_test.sh",
     "test-watch": "jest --watch",
-    "test-update": "jest -u"
+    "test-update": "jest -u",
+		"test-coverage": "jest --coverage"
   },
   "author": "mike@msanroman.io",
   "jest": {
@@ -21,7 +22,9 @@
     }
   },
   "dependencies": {
-		"@bufferapp/analyze-shared-components": "0.0.1"
+		"@bufferapp/analyze-shared-components": "0.0.1",
+		"@bufferapp/analyze-profile-selector": "0.0.1",
+		"@bufferapp/analyze-date-picker": "0.0.1"
   },
   "devDependencies": {
     "@storybook/addon-storyshots": "3.1.4",
